requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
boto3==1.26.128
botocore==1.29.128
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
dj-database-url==2.0.0
Django==4.2
django-browser-reload==1.8.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-tailwind==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
nnfs==0.5.1
numpy==1.24.3
openai==0.27.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uuid==1.30
whitenoise==6.4.0
yarl==1.9.2
